backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "cryptography>=46.0.1",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
